requirements.txt
# Core Framework
streamlit==1.30.0
langgraph==0.2.24
langchain==0.3.0
langchain-anthropic==0.2.0
langchain-community==0.3.0

# APIs
anthropic==0.34.0
tavily-python==0.5.0

# Utilities
python-dotenv==1.0.0
pydantic==2.7.0
pydantic-settings==2.3.0

# Development
pytest==8.1.0
pytest-asyncio==0.23.0
black==24.3.0
ruff==0.4.0
